Move trait VhostUserBackend to dedicated file
Move the VhostUserBackend trait to a dedicated file, preparing for
coming changes.
